+++ t/1144-isf-public-tested-by-metadata-audit.t
@@ -234,6 +234,7 @@ my $expected_tested_by = [
     't/1339-isf-scalar-storage-actor-constant-widths.t',
     't/1354-isf-scalar-storage-package-constant-widths.t',
     't/1340-isf-bank-storage-actor-constant-widths.t',
+    't/1355-isf-bank-storage-package-constant-widths.t',
     't/1341-isf-bank-storage-actor-constant-depths.t',
     't/1342-isf-transaction-port-actor-constant-widths.t',
     't/1343-isf-data-op-static-width-sources.t',